Split UseProperAssertMethodsAnalyzer into partial-class files by category
Refactor the 1567-line UseProperAssertMethodsAnalyzer into focused partial-class files grouped by assertion category, matching the pattern used in Assert source files (Assert.AreEqual.cs, Assert.Contains.cs, etc.).

The main file retains the DiagnosticAnalyzer registration, diagnostic descriptor, code-fix mode constants, top-level dispatcher, the two orchestrator methods (AnalyzeIsTrueOrIsFalseInvocation, AnalyzeAreEqualOrAreNotEqualInvocation), and shared helpers.

Category-specific recognizer helpers and their enums are moved to:

- UseProperAssertMethodsAnalyzer.NullChecks.cs (NullCheckStatus, Is*Null* helpers, RecognizeNullCheck)

- UseProperAssertMethodsAnalyzer.Equality.cs (EqualityCheckStatus, RecognizeEqualityCheck)

- UseProperAssertMethodsAnalyzer.Collection.cs (CollectionCheckStatus/CountCheckStatus/LinqPredicateCheckStatus, IsBCLCollectionType, RecognizeCollectionMethodCheck, RecognizeLinqPredicateCheck, RecognizeCountCheck overloads, TryMatch*/TryGetLinq* helpers)

- UseProperAssertMethodsAnalyzer.Comparison.cs (ComparisonCheckStatus, RecognizeComparisonCheck and IComparable helpers)

- UseProperAssertMethodsAnalyzer.String.cs (StringMethodCheckStatus, RecognizeStringMethodCheck)

No behavior change: all 237 UseProperAssertMethodsAnalyzerTests pass.
