+++ src/Analyzers/MSTest.Analyzers/UseProperAssertMethodsAnalyzer.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Immutable;
@@ -119,62 +119,8 @@ namespace MSTest.Analyzers;
 /// </list>
 /// </remarks>
 [DiagnosticAnalyzer(LanguageNames.CSharp, LanguageNames.VisualBasic)]
-public sealed class UseProperAssertMethodsAnalyzer : DiagnosticAnalyzer
+public sealed partial class UseProperAssertMethodsAnalyzer : DiagnosticAnalyzer
 {
-    private enum NullCheckStatus
-    {
-        Unknown,
-        IsNull,
-        IsNotNull,
-    }
-
-    private enum EqualityCheckStatus
-    {
-        Unknown,
-        Equals,
-        NotEquals,
-    }
-
-    private enum StringMethodCheckStatus
-    {
-        Unknown,
-        StartsWith,
-        EndsWith,
-        Contains,
-    }
-
-    private enum ComparisonCheckStatus
-    {
-        Unknown,
-        GreaterThan,
-        GreaterThanOrEqual,
-        LessThan,
-        LessThanOrEqual,
-    }
-
-    private enum CollectionCheckStatus
-    {
-        Unknown,
-        Contains,
-        ContainsWithComparer,
-    }
-
-    private enum CountCheckStatus
-    {
-        Unknown,
-        IsEmpty,
-        HasCount,
-    }
-
-    private enum LinqPredicateCheckStatus
-    {
-        Unknown,
-        Any,
-        Count,
-        WhereAny,
-        WhereCount,
-    }
-
     internal const string ProperAssertMethodNameKey = nameof(ProperAssertMethodNameKey);
 
     /// <summary>
@@ -355,392 +301,6 @@ public sealed class UseProperAssertMethodsAnalyzer : DiagnosticAnalyzer
                 AnalyzeAreEqualOrAreNotEqualInvocation(context, firstArgument, isAreEqualInvocation: false, objectTypeSymbol, enumerableTypeSymbol);
                 break;
         }
-    }
-
-    private static bool IsIsNullPattern(IOperation operation, [NotNullWhen(true)] out SyntaxNode? expressionUnderTest, out ITypeSymbol? typeOfExpressionUnderTest)
-    {
-        if (operation is IIsPatternOperation { Pattern: IConstantPatternOperation { Value: { } constantPatternValue } } isPatternOperation &&
-            constantPatternValue.WalkDownConversion() is ILiteralOperation { ConstantValue: { HasValue: true, Value: null } })
-        {
-            expressionUnderTest = isPatternOperation.Value.Syntax;
-            typeOfExpressionUnderTest = isPatternOperation.Value.WalkDownConversion().Type;
-            return true;
-        }
-
-        expressionUnderTest = null;
-        typeOfExpressionUnderTest = null;
-        return false;
-    }
-
-    private static bool IsIsNotNullPattern(IOperation operation, [NotNullWhen(true)] out SyntaxNode? expressionUnderTest, out ITypeSymbol? typeOfExpressionUnderTest)
-    {
-        if (operation is IIsPatternOperation { Pattern: INegatedPatternOperation { Pattern: IConstantPatternOperation { Value: { } constantPatternValue } } } isPatternOperation &&
-            constantPatternValue.WalkDownConversion() is ILiteralOperation { ConstantValue: { HasValue: true, Value: null } })
-        {
-            expressionUnderTest = isPatternOperation.Value.Syntax;
-            typeOfExpressionUnderTest = isPatternOperation.Value.WalkDownConversion().Type;
-            return true;
-        }
-
-        expressionUnderTest = null;
-        typeOfExpressionUnderTest = null;
-        return false;
-    }
-
-    private static bool IsEqualsNullBinaryOperator(IOperation operation, INamedTypeSymbol objectTypeSymbol, [NotNullWhen(true)] out SyntaxNode? expressionUnderTest, out ITypeSymbol? typeOfExpressionUnderTest)
-    {
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.Equals } binaryOperation &&
-            !IsExcludedOperator(binaryOperation.OperatorMethod, objectTypeSymbol))
-        {
-            if (binaryOperation.RightOperand.WalkDownConversion() is ILiteralOperation { ConstantValue: { HasValue: true, Value: null } })
-            {
-                expressionUnderTest = binaryOperation.LeftOperand.Syntax;
-                typeOfExpressionUnderTest = binaryOperation.LeftOperand.WalkDownConversion().Type;
-                return true;
-            }
-
-            if (binaryOperation.LeftOperand.WalkDownConversion() is ILiteralOperation { ConstantValue: { HasValue: true, Value: null } })
-            {
-                expressionUnderTest = binaryOperation.RightOperand.Syntax;
-                typeOfExpressionUnderTest = binaryOperation.RightOperand.WalkDownConversion().Type;
-                return true;
-            }
-        }
-
-        expressionUnderTest = null;
-        typeOfExpressionUnderTest = null;
-        return false;
-    }
-
-    private static bool IsNotEqualsNullBinaryOperator(IOperation operation, INamedTypeSymbol objectTypeSymbol, [NotNullWhen(true)] out SyntaxNode? expressionUnderTest, out ITypeSymbol? typeOfExpressionUnderTest)
-    {
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.NotEquals } binaryOperation &&
-            !IsExcludedOperator(binaryOperation.OperatorMethod, objectTypeSymbol))
-        {
-            if (binaryOperation.RightOperand.WalkDownConversion() is ILiteralOperation { ConstantValue: { HasValue: true, Value: null } })
-            {
-                expressionUnderTest = binaryOperation.LeftOperand.Syntax;
-                typeOfExpressionUnderTest = binaryOperation.LeftOperand.WalkDownConversion().Type;
-                return true;
-            }
-
-            if (binaryOperation.LeftOperand.WalkDownConversion() is ILiteralOperation { ConstantValue: { HasValue: true, Value: null } })
-            {
-                expressionUnderTest = binaryOperation.RightOperand.Syntax;
-                typeOfExpressionUnderTest = binaryOperation.RightOperand.WalkDownConversion().Type;
-                return true;
-            }
-        }
-
-        expressionUnderTest = null;
-        typeOfExpressionUnderTest = null;
-        return false;
-    }
-
-    private static NullCheckStatus RecognizeNullCheck(
-        IOperation operation,
-        INamedTypeSymbol objectTypeSymbol,
-        // Note that expressionUnderTest is guaranteed to be non-null when the method returns a value other than NullCheckStatus.Unknown.
-        // Given the current nullability attributes, there is no way to express this.
-        out SyntaxNode? expressionUnderTest,
-        out ITypeSymbol? typeOfExpressionUnderTest)
-    {
-        if (IsIsNullPattern(operation, out expressionUnderTest, out typeOfExpressionUnderTest) ||
-            IsEqualsNullBinaryOperator(operation, objectTypeSymbol, out expressionUnderTest, out typeOfExpressionUnderTest))
-        {
-            return NullCheckStatus.IsNull;
-        }
-        else if (IsIsNotNullPattern(operation, out expressionUnderTest, out typeOfExpressionUnderTest) ||
-            IsNotEqualsNullBinaryOperator(operation, objectTypeSymbol, out expressionUnderTest, out typeOfExpressionUnderTest))
-        {
-            return NullCheckStatus.IsNotNull;
-        }
-
-        return NullCheckStatus.Unknown;
-    }
-
-    private static EqualityCheckStatus RecognizeEqualityCheck(
-        IOperation operation,
-        INamedTypeSymbol objectTypeSymbol,
-        out SyntaxNode? toBecomeExpected,
-        out SyntaxNode? toBecomeActual,
-        out ITypeSymbol? leftType,
-        out ITypeSymbol? rightType)
-    {
-        if (operation is IIsPatternOperation { Pattern: IConstantPatternOperation constantPattern1 } isPattern1)
-        {
-            toBecomeExpected = constantPattern1.Syntax;
-            toBecomeActual = isPattern1.Value.Syntax;
-            leftType = constantPattern1.WalkDownConversion().Type;
-            rightType = isPattern1.Value.WalkDownConversion().Type;
-            return EqualityCheckStatus.Equals;
-        }
-        else if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.Equals } binaryOperation1 &&
-            !IsExcludedOperator(binaryOperation1.OperatorMethod, objectTypeSymbol))
-        {
-            // This is quite arbitrary. We can do extra checks to see which one (if any) looks like a "constant" and make it the expected.
-            toBecomeExpected = binaryOperation1.RightOperand.Syntax;
-            toBecomeActual = binaryOperation1.LeftOperand.Syntax;
-            leftType = binaryOperation1.RightOperand.WalkDownConversion().Type;
-            rightType = binaryOperation1.LeftOperand.WalkDownConversion().Type;
-            return EqualityCheckStatus.Equals;
-        }
-        else if (operation is IIsPatternOperation { Pattern: INegatedPatternOperation { Pattern: IConstantPatternOperation constantPattern2 } } isPattern2)
-        {
-            toBecomeExpected = constantPattern2.Syntax;
-            toBecomeActual = isPattern2.Value.Syntax;
-            leftType = constantPattern2.WalkDownConversion().Type;
-            rightType = isPattern2.Value.WalkDownConversion().Type;
-            return EqualityCheckStatus.NotEquals;
-        }
-        else if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.NotEquals } binaryOperation2 &&
-            !IsExcludedOperator(binaryOperation2.OperatorMethod, objectTypeSymbol))
-        {
-            // This is quite arbitrary. We can do extra checks to see which one (if any) looks like a "constant" and make it the expected.
-            toBecomeExpected = binaryOperation2.RightOperand.Syntax;
-            toBecomeActual = binaryOperation2.LeftOperand.Syntax;
-            leftType = binaryOperation2.RightOperand.WalkDownConversion().Type;
-            rightType = binaryOperation2.LeftOperand.WalkDownConversion().Type;
-            return EqualityCheckStatus.NotEquals;
-        }
-
-        toBecomeExpected = null;
-        toBecomeActual = null;
-        leftType = null;
-        rightType = null;
-        return EqualityCheckStatus.Unknown;
-    }
-
-    private static bool CanUseTypeAsObject(Compilation compilation, ITypeSymbol? type)
-        => type is null ||
-            compilation.ClassifyCommonConversion(type, compilation.GetSpecialType(SpecialType.System_Object)).Exists;
-
-    private static StringMethodCheckStatus RecognizeStringMethodCheck(
-        IOperation operation,
-        out SyntaxNode? stringExpression,
-        out SyntaxNode? substringExpression)
-    {
-        if (operation is IInvocationOperation invocation &&
-            invocation.TargetMethod.ContainingType?.SpecialType == SpecialType.System_String &&
-            invocation.Arguments.Length == 1)
-        {
-            string methodName = invocation.TargetMethod.Name;
-            if (methodName is "StartsWith" or "EndsWith" or "Contains" &&
-                invocation.Arguments.Length > 0 &&
-                invocation.Arguments[0].Parameter?.Type.SpecialType == SpecialType.System_String)
-            {
-                stringExpression = invocation.Instance?.Syntax;
-                substringExpression = invocation.Arguments[0].Value.Syntax;
-
-                return methodName switch
-                {
-                    "StartsWith" => StringMethodCheckStatus.StartsWith,
-                    "EndsWith" => StringMethodCheckStatus.EndsWith,
-                    "Contains" => StringMethodCheckStatus.Contains,
-                    _ => StringMethodCheckStatus.Unknown,
-                };
-            }
-        }
-
-        stringExpression = null;
-        substringExpression = null;
-        return StringMethodCheckStatus.Unknown;
-    }
-
-    private static CollectionCheckStatus RecognizeCollectionMethodCheck(
-        IOperation operation,
-        INamedTypeSymbol objectTypeSymbol,
-        INamedTypeSymbol? enumerableTypeSymbol,
-        out SyntaxNode? collectionExpression,
-        out SyntaxNode? itemExpression,
-        out SyntaxNode? comparerExpression)
-    {
-        if (operation is IInvocationOperation invocation)
-        {
-            string methodName = invocation.TargetMethod.Name;
-
-            // Check for Collection.Contains(item)
-            // We need to also ensure that invocation.TargetMethod.OriginalDefinition.Parameters[0] matches the type of the ienumerable returned by invocation.TargetMethod.ContainingType.OriginalDefinition.AllInterfaces
-            if (methodName == "Contains" &&
-                invocation.Arguments.Length == 1 &&
-                invocation.TargetMethod.OriginalDefinition.Parameters.Length == 1 &&
-                invocation.TargetMethod.OriginalDefinition.Parameters[0].Type is { } containsParameterType)
-            {
-                if (IsBCLCollectionType(invocation.TargetMethod.ContainingType, objectTypeSymbol))
-                {
-                    ITypeSymbol? enumerableElementType = invocation.TargetMethod.ContainingType.OriginalDefinition.AllInterfaces.FirstOrDefault(
-                        i => i.OriginalDefinition.SpecialType == SpecialType.System_Collections_Generic_IEnumerable_T)?.TypeArguments[0];
-
-                    if (enumerableElementType is null || enumerableElementType.Equals(containsParameterType, SymbolEqualityComparer.Default))
-                    {
-                        // If enumerableElementType is null, we expect that this is a non-generic IEnumerable. So we simply report the diagnostic.
-                        // If it's not null, ensure that the "T" of "IEnumerable<T>" matches the type of the Contains method.
-                        // The type comparison here is not for "substituted" types.
-                        // So, when analyzing KeyedCollection<TKey, TItem>.Contains(TKey), we will have:
-                        // 1. containsParameterType as the symbol referring to "TKey".
-                        // 2. enumerableElementType as the symbol referring to "TItem".
-                        // So, even if we are dealing with KeyedCollection<string, string>, the types won't match, and we won't produce a diagnostic.
-                        collectionExpression = invocation.Instance?.Syntax;
-                        itemExpression = invocation.Arguments[0].Value.Syntax;
-                        comparerExpression = null;
-                        return CollectionCheckStatus.Contains;
-                    }
-                }
-            }
-
-            // Handle LINQ Enumerable.Contains<TSource>(this IEnumerable<TSource>, TSource)
-            // In the Roslyn operation model, LINQ extension calls appear with ContainingType == Enumerable
-            // and Arguments includes the 'this' parameter, so Arguments.Length == 2.
-            if (methodName == "Contains" &&
-                invocation.Arguments.Length == 2 &&
-                enumerableTypeSymbol is not null &&
-                SymbolEqualityComparer.Default.Equals(invocation.TargetMethod.ContainingType, enumerableTypeSymbol))
-            {
-                collectionExpression = invocation.Arguments[0].Value.Syntax;
-                itemExpression = invocation.Arguments[1].Value.Syntax;
-                comparerExpression = null;
-                return CollectionCheckStatus.Contains;
-            }
-
-            // Handle LINQ Enumerable.Contains<TSource>(this IEnumerable<TSource>, TSource, IEqualityComparer<TSource>)
-            // In the Roslyn operation model, Arguments includes the 'this' parameter, so Arguments.Length == 3.
-            if (methodName == "Contains" &&
-                invocation.Arguments.Length == 3 &&
-                enumerableTypeSymbol is not null &&
-                SymbolEqualityComparer.Default.Equals(invocation.TargetMethod.ContainingType, enumerableTypeSymbol))
-            {
-                collectionExpression = invocation.Arguments[0].Value.Syntax;
-                itemExpression = invocation.Arguments[1].Value.Syntax;
-                comparerExpression = invocation.Arguments[2].Value.Syntax;
-                return CollectionCheckStatus.ContainsWithComparer;
-            }
-        }
-
-        collectionExpression = null;
-        itemExpression = null;
-        comparerExpression = null;
-        return CollectionCheckStatus.Unknown;
-    }
-
-    private static bool IsExcludedOperator(IMethodSymbol? operatorSymbol, INamedTypeSymbol objectTypeSymbol)
-        // We exclude user-defined operators from analysis. But only if they are "really" user-defined (not from BCL)
-        => operatorSymbol?.MethodKind == MethodKind.UserDefinedOperator && !IsBCLSymbol(operatorSymbol, objectTypeSymbol);
-
-    private static bool IsBCLSymbol(ISymbol symbol, INamedTypeSymbol objectTypeSymbol)
-        // object is coming from BCL and it's expected to always have a public key.
-        => symbol.ContainingAssembly.Identity.HasPublicKey == objectTypeSymbol.ContainingAssembly.Identity.HasPublicKey &&
-            symbol.ContainingAssembly.Identity.PublicKey.SequenceEqual(objectTypeSymbol.ContainingAssembly.Identity.PublicKey);
-
-    private static bool IsBCLCollectionType(ITypeSymbol type, INamedTypeSymbol objectTypeSymbol)
-        // Check if the type implements IEnumerable (but is not string)
-        => type.SpecialType != SpecialType.System_String && type.AllInterfaces.Any(i =>
-            i.OriginalDefinition.SpecialType == SpecialType.System_Collections_IEnumerable) &&
-            IsBCLSymbol(type, objectTypeSymbol);
-
-    private static ComparisonCheckStatus RecognizeComparisonCheck(
-        IOperation operation,
-        INamedTypeSymbol objectTypeSymbol,
-        INamedTypeSymbol? iComparableOfTSymbol,
-        out SyntaxNode? leftExpression,
-        out SyntaxNode? rightExpression)
-    {
-        if (operation is IBinaryOperation binaryOperation &&
-            !IsExcludedOperator(binaryOperation.OperatorMethod, objectTypeSymbol))
-        {
-            leftExpression = binaryOperation.LeftOperand.Syntax;
-            rightExpression = binaryOperation.RightOperand.Syntax;
-
-            return CanUseTypesWithComparableAsserts(iComparableOfTSymbol, binaryOperation.LeftOperand.Type, binaryOperation.RightOperand.Type)
-                ? binaryOperation.OperatorKind switch
-                {
-                    BinaryOperatorKind.GreaterThan => ComparisonCheckStatus.GreaterThan,
-                    BinaryOperatorKind.GreaterThanOrEqual => ComparisonCheckStatus.GreaterThanOrEqual,
-                    BinaryOperatorKind.LessThan => ComparisonCheckStatus.LessThan,
-                    BinaryOperatorKind.LessThanOrEqual => ComparisonCheckStatus.LessThanOrEqual,
-                    _ => ComparisonCheckStatus.Unknown,
-                }
-                : ComparisonCheckStatus.Unknown;
-        }
-
-        leftExpression = null;
-        rightExpression = null;
-        return ComparisonCheckStatus.Unknown;
-    }
-
-    private static bool CanUseTypesWithComparableAsserts(INamedTypeSymbol? iComparableOfTSymbol, ITypeSymbol? leftType, ITypeSymbol? rightType)
-        => iComparableOfTSymbol is not null
-            && leftType is not null
-            && rightType is not null
-            && SymbolEqualityComparer.Default.Equals(leftType, rightType)
-            && ImplementsIComparableOfSelf(iComparableOfTSymbol, leftType);
-
-    private static bool ImplementsIComparableOfSelf(INamedTypeSymbol iComparableOfTSymbol, ITypeSymbol type)
-    {
-        INamedTypeSymbol comparableOfSelf = iComparableOfTSymbol.Construct(type);
-        return type.AllInterfaces.Any(interfaceType => SymbolEqualityComparer.Default.Equals(interfaceType, comparableOfSelf));
-    }
-
-    private static LinqPredicateCheckStatus RecognizeLinqPredicateCheck(
-        IOperation operation,
-        INamedTypeSymbol? enumerableTypeSymbol,
-        out SyntaxNode? collectionExpression,
-        out SyntaxNode? predicateExpression,
-        out IOperation? countOperation)
-    {
-        collectionExpression = null;
-        predicateExpression = null;
-        countOperation = null;
-
-        if (enumerableTypeSymbol is null ||
-            operation is not IInvocationOperation invocation)
-        {
-            return LinqPredicateCheckStatus.Unknown;
-        }
-
-        string methodName = invocation.TargetMethod.Name;
-
-        if (!SymbolEqualityComparer.Default.Equals(invocation.TargetMethod.ContainingType, enumerableTypeSymbol))
-        {
-            return LinqPredicateCheckStatus.Unknown;
-        }
-
-        // Check for Where().Method() patterns
-        if (invocation.Arguments.Length == 1)
-        {
-            if (TryMatchWherePattern(invocation, "Any", enumerableTypeSymbol, out collectionExpression, out predicateExpression))
-            {
-                return LinqPredicateCheckStatus.WhereAny;
-            }
-
-            if (TryMatchWherePattern(invocation, "Count", enumerableTypeSymbol, out collectionExpression, out predicateExpression))
-            {
-                countOperation = operation;
-                return LinqPredicateCheckStatus.WhereCount;
-            }
-        }
-
-        // Check for direct Method(predicate) patterns
-        switch (methodName)
-        {
-            case "Any":
-                if (TryMatchLinqMethod(invocation, "Any", enumerableTypeSymbol, out collectionExpression, out predicateExpression))
-                {
-                    return LinqPredicateCheckStatus.Any;
-                }
-
-                break;
-
-            case "Count":
-                if (TryMatchLinqMethod(invocation, "Count", enumerableTypeSymbol, out collectionExpression, out predicateExpression))
-                {
-                    countOperation = operation;
-                    return LinqPredicateCheckStatus.Count;
-                }
-
-                break;
-        }
-
-        return LinqPredicateCheckStatus.Unknown;
     }
 
     private static void AnalyzeIsTrueOrIsFalseInvocation(OperationAnalysisContext context, IOperation conditionArgument, bool isTrueInvocation, INamedTypeSymbol objectTypeSymbol, INamedTypeSymbol? enumerableTypeSymbol, INamedTypeSymbol? iComparableOfTSymbol)
@@ -1221,331 +781,18 @@ public sealed class UseProperAssertMethodsAnalyzer : DiagnosticAnalyzer
         }
     }
 
-    private static bool TryMatchWherePattern(
-        IInvocationOperation invocation,
-        string methodName,
-        INamedTypeSymbol enumerableTypeSymbol,
-        out SyntaxNode? collectionExpression,
-        out SyntaxNode? predicateExpression)
-    {
-        if (invocation.TargetMethod.Name == methodName &&
-            SymbolEqualityComparer.Default.Equals(invocation.TargetMethod.ContainingType, enumerableTypeSymbol) &&
-            invocation.Arguments.Length == 1 &&
-            invocation.Arguments[0].Value is IInvocationOperation whereInvocation &&
-            whereInvocation.TargetMethod.Name == "Where" &&
-            SymbolEqualityComparer.Default.Equals(whereInvocation.TargetMethod.ContainingType, enumerableTypeSymbol) &&
-            whereInvocation.Arguments.Length == 2)
-        {
-            collectionExpression = whereInvocation.Arguments[0].Value.Syntax;
-            predicateExpression = whereInvocation.Arguments[1].Value.Syntax;
-            return true;
-        }
+    private static bool CanUseTypeAsObject(Compilation compilation, ITypeSymbol? type)
+        => type is null ||
+            compilation.ClassifyCommonConversion(type, compilation.GetSpecialType(SpecialType.System_Object)).Exists;
 
-        collectionExpression = null;
-        predicateExpression = null;
-        return false;
-    }
+    private static bool IsExcludedOperator(IMethodSymbol? operatorSymbol, INamedTypeSymbol objectTypeSymbol)
+        // We exclude user-defined operators from analysis. But only if they are "really" user-defined (not from BCL)
+        => operatorSymbol?.MethodKind == MethodKind.UserDefinedOperator && !IsBCLSymbol(operatorSymbol, objectTypeSymbol);
 
-    private static bool TryMatchLinqMethod(
-        IInvocationOperation invocation,
-        string methodName,
-        INamedTypeSymbol enumerableTypeSymbol,
-        out SyntaxNode? collectionExpression,
-        out SyntaxNode? predicateExpression)
-    {
-        if (invocation.TargetMethod.Name == methodName &&
-            SymbolEqualityComparer.Default.Equals(invocation.TargetMethod.ContainingType, enumerableTypeSymbol))
-        {
-            // Extension method with predicate: Method(collection, predicate)
-            if (invocation.Arguments.Length == 2)
-            {
-                collectionExpression = invocation.Arguments[0].Value.Syntax;
-                predicateExpression = invocation.Arguments[1].Value.Syntax;
-                return true;
-            }
-
-            // Instance method or extension without predicate: Method(collection)
-            else if (invocation.Arguments.Length == 1)
-            {
-                collectionExpression = invocation.Instance?.Syntax ?? invocation.Arguments[0].Value.Syntax;
-                predicateExpression = null;
-                return true;
-            }
-        }
-
-        collectionExpression = null;
-        predicateExpression = null;
-        return false;
-    }
-
-    private static CountCheckStatus RecognizeCountCheck(
-        IOperation operation,
-        INamedTypeSymbol objectTypeSymbol,
-        INamedTypeSymbol? enumerableTypeSymbol,
-        out SyntaxNode? collectionExpression)
-    {
-        collectionExpression = null;
-
-        // Check for collection.Count > 0 or collection.Length > 0
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.GreaterThan, LeftOperand: IPropertyReferenceOperation propertyRef, RightOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } } } &&
-            TryGetCollectionExpressionIfBCLCollectionLengthOrCount(propertyRef, objectTypeSymbol) is { } expression)
-        {
-            collectionExpression = expression;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for enumerable.Count() > 0
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.GreaterThan, LeftOperand: IInvocationOperation linqCountInv1, RightOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } } } &&
-            TryGetLinqCountNoPredicate(linqCountInv1, enumerableTypeSymbol, out SyntaxNode? linqExpr1))
-        {
-            collectionExpression = linqExpr1;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for 0 < collection.Count or 0 < collection.Length
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.LessThan, LeftOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } }, RightOperand: IPropertyReferenceOperation propertyRef2 } &&
-            TryGetCollectionExpressionIfBCLCollectionLengthOrCount(propertyRef2, objectTypeSymbol) is { } expression2)
-        {
-            collectionExpression = expression2;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for 0 < enumerable.Count()
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.LessThan, LeftOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } }, RightOperand: IInvocationOperation linqCountInv2 } &&
-            TryGetLinqCountNoPredicate(linqCountInv2, enumerableTypeSymbol, out SyntaxNode? linqExpr2))
-        {
-            collectionExpression = linqExpr2;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for collection.Count != 0 or collection.Length != 0
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.NotEquals, LeftOperand: IPropertyReferenceOperation propertyRef3, RightOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } } } &&
-            TryGetCollectionExpressionIfBCLCollectionLengthOrCount(propertyRef3, objectTypeSymbol) is { } expression3)
-        {
-            collectionExpression = expression3;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for enumerable.Count() != 0
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.NotEquals, LeftOperand: IInvocationOperation linqCountInv3, RightOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } } } &&
-            TryGetLinqCountNoPredicate(linqCountInv3, enumerableTypeSymbol, out SyntaxNode? linqExpr3))
-        {
-            collectionExpression = linqExpr3;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for 0 != collection.Count or 0 != collection.Length (reverse order)
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.NotEquals, LeftOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } }, RightOperand: IPropertyReferenceOperation propertyRef4 } &&
-            TryGetCollectionExpressionIfBCLCollectionLengthOrCount(propertyRef4, objectTypeSymbol) is { } expression4)
-        {
-            collectionExpression = expression4;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for 0 != enumerable.Count()
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.NotEquals, LeftOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } }, RightOperand: IInvocationOperation linqCountInv4 } &&
-            TryGetLinqCountNoPredicate(linqCountInv4, enumerableTypeSymbol, out SyntaxNode? linqExpr4))
-        {
-            collectionExpression = linqExpr4;
-            return CountCheckStatus.HasCount;
-        }
-
-        // Check for collection.Count == 0 or collection.Length == 0
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.Equals, LeftOperand: IPropertyReferenceOperation propertyRef5, RightOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } } } &&
-            TryGetCollectionExpressionIfBCLCollectionLengthOrCount(propertyRef5, objectTypeSymbol) is { } expression5)
-        {
-            collectionExpression = expression5;
-            return CountCheckStatus.IsEmpty;
-        }
-
-        // Check for enumerable.Count() == 0
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.Equals, LeftOperand: IInvocationOperation linqCountInv5, RightOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } } } &&
-            TryGetLinqCountNoPredicate(linqCountInv5, enumerableTypeSymbol, out SyntaxNode? linqExpr5))
-        {
-            collectionExpression = linqExpr5;
-            return CountCheckStatus.IsEmpty;
-        }
-
-        // Check for 0 == collection.Count or 0 == collection.Length (reverse order)
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.Equals, LeftOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } }, RightOperand: IPropertyReferenceOperation propertyRef6 } &&
-            TryGetCollectionExpressionIfBCLCollectionLengthOrCount(propertyRef6, objectTypeSymbol) is { } expression6)
-        {
-            collectionExpression = expression6;
-            return CountCheckStatus.IsEmpty;
-        }
-
-        // Check for 0 == enumerable.Count()
-        if (operation is IBinaryOperation { OperatorKind: BinaryOperatorKind.Equals, LeftOperand: ILiteralOperation { ConstantValue: { HasValue: true, Value: 0 } }, RightOperand: IInvocationOperation linqCountInv6 } &&
-            TryGetLinqCountNoPredicate(linqCountInv6, enumerableTypeSymbol, out SyntaxNode? linqExpr6))
-        {
-            collectionExpression = linqExpr6;
-            return CountCheckStatus.IsEmpty;
-        }
-
-        // Check for enumerable.Any() (no predicate) - direct invocation.
-        // NOTE: We return HasCount here because the caller (AnalyzeIsTrueOrIsFalseInvocation)
-        // maps HasCount → IsNotEmpty (for IsTrue) and HasCount → IsEmpty (for IsFalse),
-        // which is the correct behavior for Any(). This method is NOT called from the
-        // AreEqual/AreNotEqual path (which uses the two-argument overload), so the HasCount
-        // value won't be misinterpreted as suggesting Assert.HasCount for Any().
-        if (TryGetLinqAnyNoPredicate(operation, enumerableTypeSymbol, out SyntaxNode? linqAnyExpr))
-        {
-            collectionExpression = linqAnyExpr;
-            return CountCheckStatus.HasCount;
-        }
-
-        return CountCheckStatus.Unknown;
-    }
-
-    private static CountCheckStatus RecognizeCountCheck(
-        IOperation expectedArgument,
-        IOperation actualArgument,
-        INamedTypeSymbol objectTypeSymbol,
-        INamedTypeSymbol? enumerableTypeSymbol,
-        out SyntaxNode? nodeToBeReplaced1,
-        out SyntaxNode? replacement1,
-        out SyntaxNode? nodeToBeReplaced2,
-        out SyntaxNode? replacement2)
-    {
-        // Check if actualArgument is a count/length property
-        if (actualArgument is IPropertyReferenceOperation propertyRef &&
-            TryGetCollectionExpressionIfBCLCollectionLengthOrCount(propertyRef, objectTypeSymbol) is { } expression)
-        {
-            bool isEmpty = expectedArgument.ConstantValue.HasValue &&
-                expectedArgument.ConstantValue.Value is int expectedValue &&
-                expectedValue == 0;
-
-            if (isEmpty)
-            {
-                // We have Assert.AreEqual(0, collection.Count/Length)
-                // We want Assert.IsEmpty(collection)
-                // So, only a single replacement is needed. We replace collection.Count with collection.
-                nodeToBeReplaced1 = actualArgument.Syntax; // collection.Count
-                replacement1 = expression; // collection
-                nodeToBeReplaced2 = expectedArgument.Syntax; // 0
-                replacement2 = null;
-                return CountCheckStatus.IsEmpty;
-            }
-            else
-            {
-                // We have Assert.AreEqual(expectedCount, collection.Count/Length)
-                // We want Assert.HasCount(expectedCount, collection)
-                // Assert.HasCount takes int, so skip if expectedCount is not an int (e.g. int?, long, uint, decimal).
-                if (expectedArgument.Type?.SpecialType != SpecialType.System_Int32)
-                {
-                    nodeToBeReplaced1 = null;
-                    replacement1 = null;
-                    nodeToBeReplaced2 = null;
-                    replacement2 = null;
-                    return CountCheckStatus.Unknown;
-                }
-
-                // So, only a single replacement is needed. We replace collection.Count with collection.
-                nodeToBeReplaced1 = actualArgument.Syntax; // collection.Count
-                replacement1 = expression; // collection
-                nodeToBeReplaced2 = null;
-                replacement2 = null;
-                return CountCheckStatus.HasCount;
-            }
-        }
-
-        // Check if actualArgument is a LINQ Count() call with no predicate
-        if (actualArgument is IInvocationOperation linqCountInvocation &&
-            TryGetLinqCountNoPredicate(linqCountInvocation, enumerableTypeSymbol, out SyntaxNode? linqCollection))
-        {
-            bool isEmpty = expectedArgument.ConstantValue.HasValue &&
-                expectedArgument.ConstantValue.Value is int expectedLinqValue &&
-                expectedLinqValue == 0;
-
-            if (isEmpty)
-            {
-                // We have Assert.AreEqual(0, enumerable.Count())
-                // We want Assert.IsEmpty(enumerable)
-                nodeToBeReplaced1 = actualArgument.Syntax; // enumerable.Count()
-                replacement1 = linqCollection; // enumerable
-                nodeToBeReplaced2 = expectedArgument.Syntax; // 0
-                replacement2 = null;
-                return CountCheckStatus.IsEmpty;
-            }
-            else
-            {
-                // We have Assert.AreEqual(expectedCount, enumerable.Count())
-                // We want Assert.HasCount(expectedCount, enumerable)
-                // Assert.HasCount takes int, so skip if expectedCount is not an int (e.g. int?, long, uint, decimal).
-                if (expectedArgument.Type?.SpecialType != SpecialType.System_Int32)
-                {
-                    nodeToBeReplaced1 = null;
-                    replacement1 = null;
-                    nodeToBeReplaced2 = null;
-                    replacement2 = null;
-                    return CountCheckStatus.Unknown;
-                }
-
-                nodeToBeReplaced1 = actualArgument.Syntax; // enumerable.Count()
-                replacement1 = linqCollection; // enumerable
-                nodeToBeReplaced2 = null;
-                replacement2 = null;
-                return CountCheckStatus.HasCount;
-            }
-        }
-
-        nodeToBeReplaced1 = null;
-        replacement1 = null;
-        nodeToBeReplaced2 = null;
-        replacement2 = null;
-        return CountCheckStatus.Unknown;
-    }
-
-    private static SyntaxNode? TryGetCollectionExpressionIfBCLCollectionLengthOrCount(IPropertyReferenceOperation propertyReference, INamedTypeSymbol objectTypeSymbol)
-        => propertyReference.Property.Name is "Count" or "Length" &&
-            propertyReference.Instance?.Type is not null &&
-            (propertyReference.Instance.Type.TypeKind == TypeKind.Array || IsBCLCollectionType(propertyReference.Property.ContainingType, objectTypeSymbol))
-                ? propertyReference.Instance.Syntax
-                : null;
-
-    /// <summary>
-    /// Sets <paramref name="collectionExpression"/> to the collection syntax node if the operation is a LINQ <c>Count()</c> call with no predicate,
-    /// and returns <see langword="true"/>; otherwise sets it to <see langword="null"/> and returns <see langword="false"/>.
-    /// </summary>
-    private static bool TryGetLinqCountNoPredicate(IOperation operation, INamedTypeSymbol? enumerableTypeSymbol, [NotNullWhen(true)] out SyntaxNode? collectionExpression)
-    {
-        // LINQ Count() with no predicate is Enumerable.Count<TSource>(this IEnumerable<TSource>)
-        // In the Roslyn operation model, extension calls include 'this' in Arguments, so Arguments.Length == 1.
-        if (enumerableTypeSymbol is not null &&
-            operation is IInvocationOperation invocation &&
-            invocation.TargetMethod.Name == "Count" &&
-            invocation.Arguments.Length == 1 &&
-            SymbolEqualityComparer.Default.Equals(invocation.TargetMethod.ContainingType, enumerableTypeSymbol))
-        {
-            collectionExpression = invocation.Arguments[0].Value.Syntax;
-            return true;
-        }
-
-        collectionExpression = null;
-        return false;
-    }
-
-    /// <summary>
-    /// Sets <paramref name="collectionExpression"/> to the collection syntax node if the operation is a LINQ <c>Any()</c> call with no predicate,
-    /// and returns <see langword="true"/>; otherwise sets it to <see langword="null"/> and returns <see langword="false"/>.
-    /// </summary>
-    private static bool TryGetLinqAnyNoPredicate(IOperation operation, INamedTypeSymbol? enumerableTypeSymbol, [NotNullWhen(true)] out SyntaxNode? collectionExpression)
-    {
-        // LINQ Any() with no predicate is Enumerable.Any<TSource>(this IEnumerable<TSource>)
-        // In the Roslyn operation model, extension calls include 'this' in Arguments, so Arguments.Length == 1.
-        if (enumerableTypeSymbol is not null &&
-            operation is IInvocationOperation invocation &&
-            invocation.TargetMethod.Name == "Any" &&
-            invocation.Arguments.Length == 1 &&
-            SymbolEqualityComparer.Default.Equals(invocation.TargetMethod.ContainingType, enumerableTypeSymbol))
-        {
-            collectionExpression = invocation.Arguments[0].Value.Syntax;
-            return true;
-        }
-
-        collectionExpression = null;
-        return false;
-    }
+    private static bool IsBCLSymbol(ISymbol symbol, INamedTypeSymbol objectTypeSymbol)
+        // object is coming from BCL and it's expected to always have a public key.
+        => symbol.ContainingAssembly.Identity.HasPublicKey == objectTypeSymbol.ContainingAssembly.Identity.HasPublicKey &&
+            symbol.ContainingAssembly.Identity.PublicKey.SequenceEqual(objectTypeSymbol.ContainingAssembly.Identity.PublicKey);
 
     private static bool TryGetFirstArgumentValue(IInvocationOperation operation, [NotNullWhen(true)] out IOperation? argumentValue)
         => TryGetArgumentValueForParameterOrdinal(operation, 0, out argumentValue);
